refactor: extract bootstrap state machine into AppDelegate+Bootstrap.swift
